ml/requirements.txt
flask==2.3.3
pydantic==2.5.3
google-generativeai==0.5.4
langchain-community==0.0.38
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
sentence-transformers==2.2.2
rank_bm25==0.2.2
nltk==3.8.1
